logging improvements + refactoring

+++ BookLibrary-cdi/src/main/java/org/booklibrary/app/persistence/session/common/AbstractGenericEntityPersistence.java
@@ -1,27 +1,28 @@
-package org.booklibrary.app.persistence.session.beans;
+package org.booklibrary.app.persistence.session.common;
 
 import org.booklibrary.app.persistence.entity.AbstractBaseEntity;
 import org.booklibrary.app.persistence.id.EntityIdentifier;
-import org.booklibrary.app.persistence.session.GenericFacadeLocal;
-import org.booklibrary.app.persistence.session.GenericHomeLocal;
+import org.booklibrary.app.persistence.session.common.GenericFacadeLocal;
+import org.booklibrary.app.persistence.session.common.GenericHomeLocal;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import java.lang.reflect.ParameterizedType;
 import java.util.List;
 
 /**
  * Base abstract class for all entity persistence related classes.
+ *
+ * @see org.booklibrary.app.persistence.session.common.GenericHomeLocal
+ * @see org.booklibrary.app.persistence.session.common.GenericFacadeLocal
  */
 public abstract class AbstractGenericEntityPersistence<T extends AbstractBaseEntity, PK>
         implements GenericHomeLocal<T, PK>, GenericFacadeLocal<T, PK> {
 
     private Class<T> entityClass;
 
+    // will be injected in implementations
     public abstract EntityManager getEntityManager();
-
     public abstract Logger getLogger();
 
     public AbstractGenericEntityPersistence() {
@@ -38,10 +39,12 @@ public abstract class AbstractGenericEntityPersistence<T extends AbstractBaseEnt
             throw new IllegalArgumentException(errorMsg);
         }
 
-        getLogger().debug("findByPk invoked for entity of type [" + entityClass.getCanonicalName() +
+        getLogger().debug("find invoked for entity of type [" + entityClass.getCanonicalName() +
                 "] with pk [" + key + "]");
 
-        return getEntityManager().find(entityClass, key);
+        T entity = getEntityManager().find(entityClass, key);
+        getLogger().debug("result entity: {}", entity);
+        return entity;
     }
 
     @Override
@@ -57,15 +60,15 @@ public abstract class AbstractGenericEntityPersistence<T extends AbstractBaseEnt
 
     @Override
     public List<T> findAll() {
-
-        getLogger().debug("findAll invoked for entities [" + entityClass.getCanonicalName() + "]");
-        return getEntityManager().createQuery("FROM "
+        getLogger().debug("findAll invoked for entities of type: {}", entityClass.getCanonicalName());
+        List<T> resultList = getEntityManager().createQuery("FROM "
                 + entityClass.getName()).getResultList();
+        getLogger().debug("result list size: {}", resultList.size());
+        return resultList;
     }
 
     @Override
     public T save(T obj) throws Exception {
-
         getLogger().debug("persist invoked for object: {}", obj);
         getEntityManager().persist(obj);
         return obj;
@@ -73,7 +76,6 @@ public abstract class AbstractGenericEntityPersistence<T extends AbstractBaseEnt
 
     @Override
     public T update(T obj) throws Exception {
-
         getLogger().debug("update invoked for object {}", obj);
         return getEntityManager().merge(obj);
     }
@@ -85,32 +87,41 @@ public abstract class AbstractGenericEntityPersistence<T extends AbstractBaseEnt
     }
 
     @Override
-    public void remove(PK key) throws Exception {
-
+    public void removeByPk(PK key) throws Exception {
+        if (key == null) {
+            String errorMsg = "null argument passed to remove";
+            getLogger().error(errorMsg);
+            throw new IllegalArgumentException(errorMsg);
+        }
         getLogger().debug("remove invoked for entity of type: [" +
-                entityClass.getCanonicalName() + "] with pk: [" + key + "]");
-
+                entityClass.getCanonicalName() + "] with key: [" + key + "]");
         Object existEntity = getEntityManager().getReference(entityClass, key);
         getEntityManager().remove(existEntity);
+
     }
 
     @Override
-    public void remove(String uuid) throws Exception {
-        getLogger().debug("remove invoked for entity of type: [" +
-                entityClass.getCanonicalName() + "] with uuid: [" + uuid + "]");
+    public void removeByUuid(String uuid) throws Exception {
+        if (uuid == null) {
+            String errorMsg = "null argument passed to remove";
+            getLogger().error(errorMsg);
+            throw new IllegalArgumentException(errorMsg);
+        }
         PK key = (PK) new EntityIdentifier(uuid);
-        remove(key);
+        removeByPk(key);
 
     }
 
     @Override
-    public void removeAll() throws Exception {
+    public int removeAll() throws Exception {
         getLogger().debug("removeAll invoked for entities of type: {}",
                 entityClass.getCanonicalName());
         String entityName = entityClass.getSimpleName();
         String deleteQuery = "DELETE FROM " + entityName;
-        getEntityManager().createQuery(deleteQuery)
+        int resultSize = getEntityManager().createQuery(deleteQuery)
                 .executeUpdate();
         getEntityManager().flush();
+        getLogger().debug("removed row size: {}", resultSize);
+        return resultSize;
     }
 }
